requirements.txt
matplotlib==3.8.4
pandas==2.2.2
PyPDF2==3.0.1
python_docx==1.1.2
streamlit==1.35.0
wordcloud==1.9.3
